blogist/requirements.txt
django-tinymce
django-bootstrap5